+++ include/up-cpp/client/usubscription/v3/RpcClientUSubscription.h
@@ -20,6 +20,7 @@
 #include <up-cpp/transport/UTransport.h>
 
 #include "up-cpp/client/usubscription/v3/USubscription.h"
+#include "up-cpp/client/usubscription/v3/USubscriptionUUriBuilder.h"
 
 namespace uprotocol::core::usubscription::v3 {
 using v3::SubscriptionRequest;
@@ -48,64 +49,6 @@ struct RpcClientUSubscriptionOptions {
 	std::optional<google::protobuf::Any> subscription_details;
 };
 
-/// @struct uSubscriptionUUriBuilder
-/// @brief Structure to build uSubscription request URIs.
-///
-/// This structure is used to build URIs for uSubscription service. It uses the
-/// service options from uSubscription proto to set the authority name, ue_id,
-/// ue_version_major, and the notification topic resource ID in the URI.
-struct USubscriptionUUriBuilder {
-private:
-	/// URI for the uSubscription service
-	v1::UUri uri_;
-	/// Resource ID of the notification topic
-	uint32_t sink_resource_id_;
-
-public:
-	/// @brief Constructor for USubscriptionUUriBuilder.
-	USubscriptionUUriBuilder() {
-		// Get the service descriptor
-		const google::protobuf::ServiceDescriptor* service =
-		    uSubscription::descriptor();
-		const auto& service_options = service->options();
-
-		// Get the service options
-		const auto& service_name =
-		    service_options.GetExtension(uprotocol::service_name);
-		const auto& service_version_major =
-		    service_options.GetExtension(uprotocol::service_version_major);
-		const auto& service_id =
-		    service_options.GetExtension(uprotocol::service_id);
-		const auto& notification_topic =
-		    service_options.GetExtension(uprotocol::notification_topic, 0);
-
-		// Set the values in the URI
-		uri_.set_authority_name(service_name);
-		uri_.set_ue_id(service_id);
-		uri_.set_ue_version_major(service_version_major);
-		sink_resource_id_ = notification_topic.id();
-	}
-
-	/// @brief Get the URI with a specific resource ID.
-	///
-	/// @param resource_id The resource ID to set in the URI.
-	///
-	/// @return The URI with the specified resource ID.
-	v1::UUri getServiceUriWithResourceId(uint32_t resource_id) const {
-		v1::UUri uri = uri_;  // Copy the base URI
-		uri.set_resource_id(resource_id);
-		return uri;
-	}
-
-	/// @brief Get the notification URI.
-	///
-	/// @return The notification URI.
-	v1::UUri getNotificationUri() const {
-		v1::UUri uri = uri_;  // Copy the base URI
-		uri.set_resource_id(sink_resource_id_);
-		return uri;
-	}
-};
 
 /// @brief Interface for uEntities to create subscriptions.
 ///
@@ -136,52 +79,6 @@ struct RpcClientUSubscription : USubscription {
 	///
 	utils::Expected<SubscriptionResponse, v1::UStatus> subscribe(
 	    const SubscriptionRequest& subscription_request) override;
-	// void subscribe(google::protobuf::RpcController* controller,
-	// 	const ::uprotocol::core::usubscription::v3::SubscriptionRequest*
-	// request,
-	// 	::uprotocol::core::usubscription::v3::SubscriptionResponse* response,
-	// 	::google::protobuf::Closure* done) override;
-
-	/// @brief Unsubscribe from the topic and call uSubscription service to
-	/// close the subscription.
-	// void Unsubscribe(google::protobuf::RpcController* controller,
-	// 	const ::uprotocol::core::usubscription::v3::UnsubscribeRequest* request,
-	// 	::uprotocol::core::usubscription::v3::UnsubscribeResponse* response,
-	// 	::google::protobuf::Closure* done) override;
-
-	// /// @brief Fetch all subscriptions for a given topic or subscriber
-	// contained inside a [`FetchSubscriptionsRequest`] void
-	// FetchSubscriptions(google::protobuf::RpcController* controller, 	const
-	// ::uprotocol::core::usubscription::v3::FetchSubscriptionsRequest* request,
-	// 	::uprotocol::core::usubscription::v3::FetchSubscriptionsResponse*
-	// response,
-	// 	::google::protobuf::Closure* done) override;
-
-	// /// @brief Register for notifications relevant to a given topic inside a
-	// [`NotificationsRequest`]
-	// /// changing in subscription status.
-	// void RegisterForNotifications(google::protobuf::RpcController*
-	// controller, 	const
-	// ::uprotocol::core::usubscription::v3::NotificationsRequest* request,
-	// 	::uprotocol::core::usubscription::v3::NotificationsResponse* response,
-	// 	::google::protobuf::Closure* done) override;
-
-	// /// @brief Unregister for notifications relevant to a given topic inside
-	// a [`NotificationsRequest`]
-	// /// changing in subscription status.
-	// void UnregisterForNotifications(google::protobuf::RpcController*
-	// controller, 	const
-	// ::uprotocol::core::usubscription::v3::NotificationsRequest* request,
-	// 	::uprotocol::core::usubscription::v3::NotificationsResponse* response,
-	// 	::google::protobuf::Closure* done) override;
-
-	// /// @brief Fetch a list of subscribers that are currently subscribed to a
-	// given topic in a [`FetchSubscribersRequest`] void
-	// FetchSubscribers(google::protobuf::RpcController* controller, 	const
-	// ::uprotocol::core::usubscription::v3::FetchSubscribersRequest* request,
-	// 	::uprotocol::core::usubscription::v3::FetchSubscribersResponse*
-	// response,
-	// 	::google::protobuf::Closure* done) override;
 
 	/// @brief Destructor
 	~RpcClientUSubscription() override = default;
@@ -201,9 +98,6 @@ private:
 	// Transport
 	std::shared_ptr<transport::UTransport> transport_;
 
-	// Topic to subscribe to
-	const v1::UUri subscription_topic_;
-
 	// Additional details about uSubscription service
 	RpcClientUSubscriptionOptions rpc_client_usubscription_options_;
 
@@ -218,9 +112,8 @@ private:
 	//     std::shared_ptr<uprotocol::transport::UTransport>&&, const v1::UUri&&,
 	//     RpcClientUSubscriptionOptions&&);
 
-public:
 	/// @brief Build SubscriptionRequest for subscription request
-	SubscriptionRequest buildSubscriptionRequest();
+	SubscriptionRequest buildSubscriptionRequest(const v1::UUri& subscription_topic);
 	//
 	// /// @brief  Build UnsubscriptionRequest for unsubscription request
 	// UnsubscribeRequest buildUnsubscriptionRequest();
