+++ include/up-cpp/client/usubscription/v3/USubscriptionUUriBuilder.h
@@ -1,0 +1,49 @@
+//
+// Created by max on 10.05.25.
+//
+
+#ifndef USUBSCRIPTIONUURIBUILDER_H
+#define USUBSCRIPTIONUURIBUILDER_H
+
+#include <uprotocol/uoptions.pb.h>
+#include <uprotocol/core/usubscription/v3/usubscription.pb.h>
+
+#include "up-cpp/datamodel/validator/UUri.h"
+
+namespace uprotocol::core::usubscription::v3 {
+/// @struct USubscriptionUUriBuilder
+/// @brief Structure to build uSubscription request URIs.
+///
+/// This structure is used to build URIs for uSubscription service. It uses the
+/// service options from uSubscription proto to set the authority name, ue_id,
+/// ue_version_major, and the notification topic resource ID in the URI.
+struct USubscriptionUUriBuilder {
+
+	/// @brief Constructor for USubscriptionUUriBuilder.
+	explicit USubscriptionUUriBuilder();
+
+	/// @brief Get the URI with a specific resource ID.
+	///
+	/// @param resource_id The resource ID to set in the URI.
+	///
+	/// @return The URI with the specified resource ID.
+	v1::UUri getServiceUriWithResourceId(uint32_t resource_id) const;
+
+	/// @brief Get the notification URI.
+	///
+	/// @return The notification URI.
+	v1::UUri getNotificationUri() const {
+		v1::UUri uri = base_uri_;  // Copy the base URI
+		uri.set_resource_id(sink_resource_id_);
+		return uri;
+	}
+
+	private:
+		/// URI for the uSubscription service
+		v1::UUri base_uri_;
+		/// Resource ID of the notification topic
+		uint32_t sink_resource_id_;
+	};
+} // namespace uprotocol::core::usubscription::v3
+
+#endif //USUBSCRIPTIONUURIBUILDER_H
